requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-money==1.0
Pillow==7.0.0
pkg-resources==0.0.0
py-moneyed==0.8.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
